+++ example/app.babel.js
@@ -4,4 +4,4 @@ require('babel-core/register')({
 });
 require('babel-polyfill');
 
-require('./src/app.js');
+require('./app.js');